remove `{Back,Fore,Front}Fang` & add `FangAction`; next: update docs
