+++ ohkami/Cargo.toml
@@ -36,7 +36,7 @@ sha2          = { version = "0.10.8", default-features = false }
 
 
 [features]
-default       = ["testing"]
+#default       = ["testing"]
 rt_tokio      = ["dep:tokio"]
 rt_async-std  = ["dep:async-std"]
 nightly       = []
@@ -50,10 +50,10 @@ DEBUG = [
     "tokio?/rt-multi-thread",
     "async-std?/attributes",
 ]
-#default = [
-#    "testing",
-#    #"websocket",
-#    "rt_tokio",
-#    #"rt_async-std",
-#    "DEBUG",
-#]
+default = [
+    "testing",
+    #"websocket",
+    "rt_tokio",
+    #"rt_async-std",
+    "DEBUG",
+]
