+++ ohkami/src/fangs/middleware/mod.rs
@@ -1,4 +1,4 @@
-pub mod utils;
+pub mod util;
 
 use super::{Fang, BoxedFPC};
